getbsize() now takes a pointer to a size_t as its first arg.  A pointer to an int no longer works on an LP64 platform...
git-svn-id: svn+ssh://svn.freebsd.org/base/head@105882 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.sbin/pkg_install/info/show.c
@@ -233,7 +233,7 @@ show_size(const char *title, Package *plist)
     char tmp[FILENAME_MAX];
     unsigned long size = 0;
     long blksize;
-    int headerlen;
+    size_t headerlen;
     char *descr;
 
     descr = getbsize(&headerlen, &blksize);
